+++ workers/unity/Packages/io.improbable.gdk.core/Worker/ConnectionHandlers/ConnectionFlowInitializers.cs
@@ -1,12 +1,9 @@
-using System;
 using System.Collections.Generic;
-using Improbable.Worker.CInterop;
 
 namespace Improbable.Gdk.Core
 {
-    // TODO: Should this be contravarient?
     /// <summary>
-    ///    Represents an object which can initialize a connection flow of a certain type.
+    ///     Represents an object which can initialize a connection flow of a certain type.
     /// </summary>
     /// <typeparam name="TConnectionFlow">The type of connection flow that this object can initialize.</typeparam>
     public interface IConnectionFlowInitializer<TConnectionFlow> where TConnectionFlow : IConnectionFlow
@@ -19,13 +16,12 @@ namespace Improbable.Gdk.Core
     }
 
     /// <summary>
-    ///     Represents an object which can initialize the <see cref="ReceptionistFlow"/>, <see cref="LocatorFlow"/>,
-    ///     and <see cref="AlphaLocatorFlow"/> connection flows from the command line.
+    ///     Represents an object which can initialize the <see cref="ReceptionistFlow" />, <see cref="LocatorFlow" />,
+    ///     and <see cref="LocatorFlow" /> connection flows from the command line.
     /// </summary>
     public class CommandLineConnectionFlowInitializer :
         IConnectionFlowInitializer<ReceptionistFlow>,
-        IConnectionFlowInitializer<LocatorFlow>,
-        IConnectionFlowInitializer<AlphaLocatorFlow>
+        IConnectionFlowInitializer<LocatorFlow>
     {
         private readonly CommandLineArgs commandLineArgs;
 
@@ -45,17 +41,9 @@ namespace Improbable.Gdk.Core
         /// <returns>The connection service to use.</returns>
         public ConnectionService GetConnectionService()
         {
-            if (commandLineArgs.Contains(RuntimeConfigNames.SteamDeploymentTag) &&
-                commandLineArgs.Contains(RuntimeConfigNames.SteamTicket))
+            if (commandLineArgs.Contains(RuntimeConfigNames.LoginToken) && commandLineArgs.Contains(RuntimeConfigNames.PlayerIdentityToken))
             {
                 return ConnectionService.Locator;
-            }
-
-            if (commandLineArgs.Contains(RuntimeConfigNames.LoginToken))
-            {
-                return commandLineArgs.Contains(RuntimeConfigNames.PlayerIdentityToken)
-                    ? ConnectionService.AlphaLocator
-                    : ConnectionService.Locator;
             }
 
             return ConnectionService.Receptionist;
@@ -63,43 +51,24 @@ namespace Improbable.Gdk.Core
 
         public void Initialize(ReceptionistFlow receptionist)
         {
-            commandLineArgs.TryGetCommandLineValue(RuntimeConfigNames.ReceptionistHost, ref receptionist.ReceptionistHost);
-            commandLineArgs.TryGetCommandLineValue(RuntimeConfigNames.ReceptionistPort, ref receptionist.ReceptionistPort);
+            commandLineArgs.TryGetCommandLineValue(RuntimeConfigNames.ReceptionistHost,
+                ref receptionist.ReceptionistHost);
+            commandLineArgs.TryGetCommandLineValue(RuntimeConfigNames.ReceptionistPort,
+                ref receptionist.ReceptionistPort);
             commandLineArgs.TryGetCommandLineValue(RuntimeConfigNames.WorkerId, ref receptionist.WorkerId);
         }
 
         public void Initialize(LocatorFlow locator)
         {
             commandLineArgs.TryGetCommandLineValue(RuntimeConfigNames.LocatorHost, ref locator.LocatorHost);
-            commandLineArgs.TryGetCommandLineValue(RuntimeConfigNames.ProjectName,
-                ref locator.LocatorParameters.ProjectName);
-            commandLineArgs.TryGetCommandLineValue(RuntimeConfigNames.LoginToken,
-                ref locator.LocatorParameters.LoginToken.Token);
-            commandLineArgs.TryGetCommandLineValue(RuntimeConfigNames.SteamDeploymentTag,
-                ref locator.LocatorParameters.Steam.DeploymentTag);
-            commandLineArgs.TryGetCommandLineValue(RuntimeConfigNames.SteamTicket,
-                ref locator.LocatorParameters.Steam.Ticket);
+            commandLineArgs.TryGetCommandLineValue(RuntimeConfigNames.PlayerIdentityToken,
+                ref locator.PlayerIdentityToken);
+            commandLineArgs.TryGetCommandLineValue(RuntimeConfigNames.LoginToken, ref locator.LoginToken);
 
-            if (!string.IsNullOrEmpty(locator.LocatorParameters.Steam.Ticket) && !string.IsNullOrEmpty(locator.LocatorParameters.Steam.DeploymentTag))
+            if (!string.IsNullOrEmpty(locator.PlayerIdentityToken)
+                && !string.IsNullOrEmpty(locator.LoginToken))
             {
-                locator.LocatorParameters.CredentialsType = LocatorCredentialsType.Steam;
-            }
-            else if (!string.IsNullOrEmpty(locator.LocatorParameters.LoginToken.Token))
-            {
-                locator.LocatorParameters.CredentialsType = LocatorCredentialsType.LoginToken;
-            }
-        }
-
-        public void Initialize(AlphaLocatorFlow alphaLocator)
-        {
-            commandLineArgs.TryGetCommandLineValue(RuntimeConfigNames.LocatorHost, ref alphaLocator.LocatorHost);
-            commandLineArgs.TryGetCommandLineValue(RuntimeConfigNames.PlayerIdentityToken, ref alphaLocator.LocatorParameters.PlayerIdentity.PlayerIdentityToken);
-            commandLineArgs.TryGetCommandLineValue(RuntimeConfigNames.LoginToken, ref alphaLocator.LocatorParameters.PlayerIdentity.LoginToken);
-
-            if (!string.IsNullOrEmpty(alphaLocator.LocatorParameters.PlayerIdentity.PlayerIdentityToken)
-                && !string.IsNullOrEmpty(alphaLocator.LocatorParameters.PlayerIdentity.LoginToken))
-            {
-                alphaLocator.UseDevAuthFlow = false;
+                locator.UseDevAuthFlow = false;
             }
         }
     }